move block model functions to utils/models
