+++ geoapps/utils/models.py
@@ -194,3 +194,61 @@ class RectangularBlock:
             self._vertices = xyz
 
         return self._vertices
+
+
+def truncate_locs_depths(locs, depth_core):
+    zmax = locs[:, 2].max()  # top of locs
+    below_core_ind = (zmax - locs[:, 2]) > depth_core
+    core_bottom_elev = zmax - depth_core
+    locs[
+        below_core_ind, 2
+    ] = core_bottom_elev  # sets locations below core to core bottom
+    return locs
+
+
+def minimum_depth_core(locs, depth_core, core_z_cell_size):
+    zrange = locs[:, 2].max() - locs[:, 2].min()  # locs z range
+    if depth_core >= zrange:
+        return depth_core - zrange + core_z_cell_size
+    else:
+        return depth_core
+
+
+def find_top_padding(obj, core_z_cell_size):
+    pad_sum = 0
+    for h in np.abs(np.diff(obj.z_cell_delimiters)):
+        if h != core_z_cell_size:
+            pad_sum += h
+        else:
+            return pad_sum
+
+
+def get_block_model(workspace, name, locs, h, depth_core, pads, expansion_factor):
+
+    locs = truncate_locs_depths(locs, depth_core)
+    depth_core = minimum_depth_core(locs, depth_core, h[2])
+    mesh = mesh_utils.mesh_builder_xyz(
+        locs,
+        h,
+        padding_distance=[
+            [pads[0], pads[1]],
+            [pads[2], pads[3]],
+            [pads[4], pads[5]],
+        ],
+        depth_core=depth_core,
+        expansion_factor=expansion_factor,
+    )
+
+    object_out = BlockModel.create(
+        workspace,
+        origin=[mesh.x0[0], mesh.x0[1], locs[:, 2].max()],
+        u_cell_delimiters=mesh.vectorNx - mesh.x0[0],
+        v_cell_delimiters=mesh.vectorNy - mesh.x0[1],
+        z_cell_delimiters=-(mesh.x0[2] + mesh.hz.sum() - mesh.vectorNz[::-1]),
+        name=name,
+    )
+
+    top_padding = find_top_padding(object_out, h[2])
+    object_out.origin["z"] += top_padding
+
+    return object_out
